Move EvaluateTreebank to the metrics directory
