+++ src/edu/stanford/nlp/parser/dvparser/CrossValidateTestOptions.java
@@ -7,8 +7,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import edu.stanford.nlp.parser.common.ArgUtils;
-import edu.stanford.nlp.parser.lexparser.EvaluateTreebank;
 import edu.stanford.nlp.parser.lexparser.LexicalizedParser;
+import edu.stanford.nlp.parser.metrics.EvaluateTreebank;
 import edu.stanford.nlp.trees.Treebank;
 import edu.stanford.nlp.util.Pair;
