+++ src/edu/stanford/nlp/parser/dvparser/DVParser.java
@@ -20,10 +20,10 @@ import edu.stanford.nlp.math.ArrayMath;
 import edu.stanford.nlp.optimization.QNMinimizer;
 import edu.stanford.nlp.parser.common.ArgUtils;
 import edu.stanford.nlp.parser.common.ParserQuery;
-import edu.stanford.nlp.parser.lexparser.EvaluateTreebank;
 import edu.stanford.nlp.parser.lexparser.LexicalizedParser;
 import edu.stanford.nlp.parser.lexparser.Options;
 import edu.stanford.nlp.parser.lexparser.TrainOptions;
+import edu.stanford.nlp.parser.metrics.EvaluateTreebank;
 import edu.stanford.nlp.trees.CompositeTreeTransformer;
 import edu.stanford.nlp.trees.TreeTransformer;
 import edu.stanford.nlp.trees.Tree;
